Add book example chap 14 and ex 14.1
Move to latest version in order to use vc12 instead of vc10. No version
mismatch anymore.

+++ the_concepts_and_practice_of_math_fin/XLL_Project - projects b/source/cppinterface.h
@@ -73,7 +73,7 @@ _ZeroCoupon( double r //interest rate
 
 double // evaluate final price bsd on Euler stepping
 _FinalPrice_EulerStepping(double Spot //spot
-                         ,double Strike //strike
+                         ,short Strike //strike
                          ,double r //interest rate
                          ,double d //dividend yield
                          ,double Vol //volatility
